% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyHWFunctions.R
\name{myLinearRegression}
\alias{myLinearRegression}
\title{Add together two numbers.}
\usage{
myLinearRegression(Y, X, sub)
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This is truly a great and much-needed function
}
\examples{
aGreatFunction(1, 1)
aGreatFunction(10, 1)
}
