% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyHWFunctions.R
\name{myLinearRegression}
\alias{myLinearRegression}
\title{Linear Regression/Covariate Exploration}
\usage{
myLinearRegression(Y, X, sub)
}
\arguments{
\item{Y}{A vector of outcomes.}

\item{X}{a matrix of covariates.}

\item{sub:}{a list of subjects (i.e. a set of integers corresponding to rows in X).}
}
\value{
p value of \code{x}, coefficients of \code{x}, and graphs of covariates of \code{x}.
}
\description{
This package takes in a vector of outcomes, a matrix of covariates, and the number of rows of the independent variable. It then takes the linear regression of the given X and Y variables and returns the corresponding p-values and coefficients. The function also returns graphs of each of the covariates.
}
\examples{
myLinearRegression(myData[,1],myData[,2:6],1:30)
myLinearRegression(myData[,1],myData[,2:4],1:40)
}
