% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MyHWFunctions.R
\name{myLinearRegression}
\alias{myLinearRegression}
\title{Takes the linear regression of X and Y variables and results in p-values, coefficients, and graphs of covariates.}
\usage{
myLinearRegression(Y, X, sub)
}
\arguments{
\item{Y}{A vector of outcomes.}

\item{X}{a matrix of covariates.}

\item{sub:}{a list of subjects (i.e. a set of integers corresponding to rows in X).}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
This is truly a great and much-needed function
}
\examples{
myLinearRegression(myData[,1],myData[,2:6],1:30)
myLinearRegression(myData[,1],myData[,2:4],1:40)
}
